Qt client: Fix crash at startup when autoconnect is on
menus_init() can be called while the ruleset isn't yet ready. A subsequent call
to menus_update() fails. Fix by initializing the menus from menus_update() if
needed.

The crash didn't happen every time but was more frequent when autoconnecting to
a running game with a modified version of the trident tileset. Stock amplio2
was not triggering the bug.